pyproject.toml
[project]
name = "agentprof"
version = "0.0.1"
description = "Local-first CLI for AI-agent failure-and-waste reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.0",
    "pydantic>=2.12.0",
    "pyyaml>=6.0.0",
    "rich>=14.0.0",
    "typer>=0.20.0",
]

[project.scripts]
agentprof = "agentprof.cli:app"

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
